requirements.txt
absl-py==2.1.0
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.8.2
blis==1.0.1
boto3==1.35.54
botocore==1.35.54
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.87
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
coverage==7.6.6
cymem==2.0.8
etils==1.10.0
execnet==2.1.1
factory_boy==3.3.1
Faker==30.8.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
flatbuffers==24.3.25
flax==0.10.1
fsspec==2024.10.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
h5py==3.12.1
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
jmespath==1.0.1
keras==2.13.1
keras-models==0.0.7
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
Mako==1.3.6
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
marshmallow-enum==1.5.1
mdurl==0.1.2
ml_dtypes==0.5.0
msgpack==1.1.0
murmurhash==1.0.10
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.0
orbax-checkpoint==0.9.1
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pep8==1.7.1
pillow==11.0.0
pluggy==1.5.0
preshed==3.0.9
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cache==1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.3
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-decision-forests==1.5.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflowjs==4.22.0
tensorstore==0.1.67
termcolor==2.5.0
tf-keras==2.15.0
thinc==8.3.2
toolz==1.0.0
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.6
wrapt==1.16.0
wurlitzer==3.1.1
zipp==3.21.0
